+++ packages/gatsby/index.d.ts
@@ -23,6 +23,8 @@ export const useStaticQuery: <TData = any>(query: any) => TData
 
 export const parsePath: (path: string) => WindowLocation
 
+export const prefetchPathname: (path: string) => void
+
 export interface PageRendererProps {
   location: WindowLocation
 }
@@ -795,7 +797,11 @@ export interface Actions {
   deleteNodes(nodes: string[], plugin?: ActionPlugin): void
 
   /** @see https://www.gatsbyjs.org/docs/actions/#createNode */
-  createNode(node: Node, plugin?: ActionPlugin, options?: ActionOptions): void
+  createNode(
+    node: NodeInput,
+    plugin?: ActionPlugin,
+    options?: ActionOptions
+  ): void
 
   /** @see https://www.gatsbyjs.org/docs/actions/#touchNode */
   touchNode(node: { nodeId: string }, plugin?: ActionPlugin): void
@@ -1114,42 +1120,25 @@ export interface ServiceWorkerArgs extends BrowserPluginArgs {
   serviceWorker: ServiceWorkerRegistration
 }
 
-export interface Node {
-  path?: string
+export interface NodeInput {
   id: string
-  parent: string
-  children: Node[]
-  fields?: Record<string, string>
+  parent?: string
+  children?: string[]
   internal: {
     type: string
-    mediaType: string
-    content: string
+    mediaType?: string
+    content?: string
     contentDigest: string
-    owner: string
     description?: string
   }
-  resolve?: string
-  name?: string
-  version?: string
-  pluginOptions?: PluginOptions
-  nodeAPIs?: any[]
-  browserAPIs?: any[]
-  ssrAPIs?: any[]
-  pluginFilepath?: string
-  packageJson?: PackageJson
-  siteMetadata?: Record<string, any>
-  port?: string
-  host?: string
-  pathPrefix?: string
-  polyfill?: boolean
-  buildTime?: string
-  jsonName?: string
-  internalComponentName?: string
-  matchPath?: unknown
-  component?: string
-  componentChunkName?: string
-  context?: Record<string, any>
-  pluginCreatorId?: string
-  componentPath?: string
+  [key: string]: unknown
+}
+
+export interface Node extends NodeInput {
+  parent: string
+  children: string[]
+  internal: NodeInput["internal"] & {
+    owner: string
+  }
   [key: string]: unknown
 }