+++ packages/gatsby-dev-cli/CHANGELOG.md
@@ -3,6 +3,10 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [2.5.5](https://github.com/gatsbyjs/gatsby/compare/gatsby-dev-cli@2.5.4...gatsby-dev-cli@2.5.5) (2019-07-13)
+
+**Note:** Version bump only for package gatsby-dev-cli
+
 ## [2.5.4](https://github.com/gatsbyjs/gatsby/compare/gatsby-dev-cli@2.5.3...gatsby-dev-cli@2.5.4) (2019-07-12)
 
 ### Features