+++ packages/gatsby-page-utils/CHANGELOG.md
@@ -3,6 +3,10 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.0.5](https://github.com/gatsbyjs/gatsby/compare/gatsby-page-utils@0.0.4...gatsby-page-utils@0.0.5) (2019-07-13)
+
+**Note:** Version bump only for package gatsby-page-utils
+
 ## [0.0.4](https://github.com/gatsbyjs/gatsby/compare/gatsby-page-utils@0.0.3...gatsby-page-utils@0.0.4) (2019-07-12)
 
 ### Bug Fixes