+++ packages/gatsby-plugin-mdx/CHANGELOG.md
@@ -3,6 +3,10 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.0.13](https://github.com/gatsbyjs/gatsby/compare/gatsby-plugin-mdx@1.0.12...gatsby-plugin-mdx@1.0.13) (2019-07-13)
+
+**Note:** Version bump only for package gatsby-plugin-mdx
+
 ## [1.0.12](https://github.com/gatsbyjs/gatsby/compare/gatsby-plugin-mdx@1.0.11...gatsby-plugin-mdx@1.0.12) (2019-07-12)
 
 ### Bug Fixes